editors - implement new text editor service and adopt (#126960)
