+++ src/vs/workbench/contrib/output/browser/logViewer.ts
@@ -22,6 +22,7 @@ import { IEditorService } from 'vs/workbench/services/editor/common/editorServic
 import { ITextFileService } from 'vs/workbench/services/textfile/common/textfiles';
 import { IFileService } from 'vs/platform/files/common/files';
 import { ILabelService } from 'vs/platform/label/common/label';
+import { IEditorResolverService } from 'vs/workbench/services/editor/common/editorResolverService';
 
 export class LogViewerInput extends TextResourceEditorInput {
 
@@ -37,7 +38,8 @@ export class LogViewerInput extends TextResourceEditorInput {
 		@ITextFileService textFileService: ITextFileService,
 		@IEditorService editorService: IEditorService,
 		@IFileService fileService: IFileService,
-		@ILabelService labelService: ILabelService
+		@ILabelService labelService: ILabelService,
+		@IEditorResolverService editorResolverService: IEditorResolverService
 	) {
 		super(
 			URI.from({ scheme: LOG_SCHEME, path: outputChannelDescriptor.id }),
@@ -49,7 +51,8 @@ export class LogViewerInput extends TextResourceEditorInput {
 			textFileService,
 			editorService,
 			fileService,
-			labelService
+			labelService,
+			editorResolverService
 		);
 	}
 }
