+++ src/vs/workbench/contrib/preferences/common/preferencesContribution.ts
@@ -22,7 +22,7 @@ import { IWorkbenchContribution } from 'vs/workbench/common/contributions';
 import { IEditorInputWithOptions } from 'vs/workbench/common/editor';
 import { SideBySideEditorInput } from 'vs/workbench/common/editor/sideBySideEditorInput';
 import { RegisteredEditorPriority, IEditorResolverService } from 'vs/workbench/services/editor/common/editorResolverService';
-import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
+import { ITextEditorService } from 'vs/workbench/services/textfile/common/textEditorService';
 import { DEFAULT_SETTINGS_EDITOR_SETTING, FOLDER_SETTINGS_PATH, IPreferencesService, USE_SPLIT_JSON_SETTING } from 'vs/workbench/services/preferences/common/preferences';
 
 const schemaRegistry = Registry.as<JSONContributionRegistry.IJSONContributionRegistry>(JSONContributionRegistry.Extensions.JSONContribution);
@@ -40,7 +40,7 @@ export class PreferencesContribution implements IWorkbenchContribution {
 		@IWorkspaceContextService private readonly workspaceService: IWorkspaceContextService,
 		@IConfigurationService private readonly configurationService: IConfigurationService,
 		@IEditorResolverService private readonly editorResolverService: IEditorResolverService,
-		@IEditorService private readonly editorService: IEditorService,
+		@ITextEditorService private readonly textEditorService: ITextEditorService
 	) {
 		this.settingsListener = this.configurationService.onDidChangeConfiguration(e => {
 			if (e.affectsConfiguration(USE_SPLIT_JSON_SETTING) || e.affectsConfiguration(DEFAULT_SETTINGS_EDITOR_SETTING)) {
@@ -94,7 +94,7 @@ export class PreferencesContribution implements IWorkbenchContribution {
 						}
 					}
 
-					return { editor: this.editorService.createEditorInput({ resource }), options };
+					return { editor: this.textEditorService.createTextEditor({ resource }), options };
 				}
 			);
 		}
