+++ src/vs/workbench/services/editor/browser/editorService.ts
@@ -5,18 +5,13 @@
 
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { IResourceEditorInput, IEditorOptions, EditorActivation, EditorResolution, IResourceEditorInputIdentifier, ITextResourceEditorInput } from 'vs/platform/editor/common/editor';
-import { SideBySideEditor, IEditorInput, IEditorPane, GroupIdentifier, IFileEditorInput, IUntitledTextResourceEditorInput, IResourceDiffEditorInput, IEditorFactoryRegistry, EditorExtensions, IEditorInputWithOptions, isEditorInputWithOptions, IEditorIdentifier, IEditorCloseEvent, ITextDiffEditorPane, IRevertOptions, SaveReason, EditorsOrder, IWorkbenchEditorConfiguration, EditorResourceAccessor, IVisibleEditorPane, EditorInputCapabilities, isResourceDiffEditorInput, IUntypedEditorInput, DEFAULT_EDITOR_ASSOCIATION, isResourceEditorInput, isEditorInput, isEditorInputWithOptionsAndGroup, isResourceSideBySideEditorInput, IUntypedFileEditorInput } from 'vs/workbench/common/editor';
-import { EditorInput } from 'vs/workbench/common/editor/editorInput';
+import { SideBySideEditor, IEditorInput, IEditorPane, GroupIdentifier, IUntitledTextResourceEditorInput, IResourceDiffEditorInput, IEditorInputWithOptions, isEditorInputWithOptions, IEditorIdentifier, IEditorCloseEvent, ITextDiffEditorPane, IRevertOptions, SaveReason, EditorsOrder, IWorkbenchEditorConfiguration, EditorResourceAccessor, IVisibleEditorPane, EditorInputCapabilities, isResourceDiffEditorInput, IUntypedEditorInput, isResourceEditorInput, isEditorInput, isEditorInputWithOptionsAndGroup } from 'vs/workbench/common/editor';
 import { SideBySideEditorInput } from 'vs/workbench/common/editor/sideBySideEditorInput';
-import { TextResourceEditorInput } from 'vs/workbench/common/editor/textResourceEditorInput';
-import { Registry } from 'vs/platform/registry/common/platform';
 import { ResourceMap } from 'vs/base/common/map';
-import { IUntitledTextEditorService } from 'vs/workbench/services/untitled/common/untitledTextEditorService';
 import { IFileService, FileOperationEvent, FileOperation, FileChangesEvent, FileChangeType } from 'vs/platform/files/common/files';
-import { Schemas } from 'vs/base/common/network';
 import { Event, Emitter, DebounceEmitter } from 'vs/base/common/event';
 import { URI } from 'vs/base/common/uri';
-import { basename, joinPath } from 'vs/base/common/resources';
+import { joinPath } from 'vs/base/common/resources';
 import { DiffEditorInput } from 'vs/workbench/common/editor/diffEditorInput';
 import { IEditorGroupsService, IEditorGroup, GroupsOrder, IEditorReplacement, GroupChangeKind, isEditorReplacement } from 'vs/workbench/services/editor/common/editorGroupsService';
 import { IUntypedEditorReplacement, IEditorService, ISaveEditorsOptions, ISaveAllEditorsOptions, IRevertAllEditorsOptions, IBaseSaveRevertAllEditorOptions, IOpenEditorsOptions, PreferredGroup, isPreferredGroup } from 'vs/workbench/services/editor/common/editorService';
@@ -28,19 +23,16 @@ import { IEditorGroupView, EditorServiceImpl } from 'vs/workbench/browser/parts/
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { isUndefined, withNullAsUndefined } from 'vs/base/common/types';
 import { EditorsObserver } from 'vs/workbench/browser/parts/editor/editorsObserver';
-import { IUntitledTextEditorModel } from 'vs/workbench/services/untitled/common/untitledTextEditorModel';
-import { UntitledTextEditorInput } from 'vs/workbench/services/untitled/common/untitledTextEditorInput';
 import { Promises, timeout } from 'vs/base/common/async';
 import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
 import { indexOfPath } from 'vs/base/common/extpath';
 import { IUriIdentityService } from 'vs/workbench/services/uriIdentity/common/uriIdentity';
-import { RegisteredEditorPriority, IEditorResolverService, ResolvedStatus } from 'vs/workbench/services/editor/common/editorResolverService';
+import { IEditorResolverService, ResolvedStatus } from 'vs/workbench/services/editor/common/editorResolverService';
 import { IWorkingCopyService } from 'vs/workbench/services/workingCopy/common/workingCopyService';
 import { IWorkspaceTrustRequestService, WorkspaceTrustUriResponse } from 'vs/platform/workspace/common/workspaceTrust';
 import { IHostService } from 'vs/workbench/services/host/browser/host';
 import { findGroup } from 'vs/workbench/services/editor/common/editorGroupFinder';
-
-type CachedEditorInput = TextResourceEditorInput | IFileEditorInput | UntitledTextEditorInput;
+import { ITextEditorService } from 'vs/workbench/services/textfile/common/textEditorService';
 
 export class EditorService extends Disposable implements EditorServiceImpl {
 
@@ -68,11 +60,8 @@ export class EditorService extends Disposable implements EditorServiceImpl {
 
 	//#endregion
 
-	private readonly fileEditorFactory = Registry.as<IEditorFactoryRegistry>(EditorExtensions.EditorFactory).getFileEditorFactory();
-
 	constructor(
 		@IEditorGroupsService private readonly editorGroupService: IEditorGroupsService,
-		@IUntitledTextEditorService private readonly untitledTextEditorService: IUntitledTextEditorService,
 		@IInstantiationService private readonly instantiationService: IInstantiationService,
 		@IFileService private readonly fileService: IFileService,
 		@IConfigurationService private readonly configurationService: IConfigurationService,
@@ -82,16 +71,13 @@ export class EditorService extends Disposable implements EditorServiceImpl {
 		@IWorkingCopyService private readonly workingCopyService: IWorkingCopyService,
 		@IWorkspaceTrustRequestService private readonly workspaceTrustRequestService: IWorkspaceTrustRequestService,
 		@IHostService private readonly hostService: IHostService,
+		@ITextEditorService private readonly textEditorService: ITextEditorService
 	) {
 		super();
 
 		this.onConfigurationUpdated(configurationService.getValue<IWorkbenchEditorConfiguration>());
 
 		this.registerListeners();
-
-		// Register the default editor to the editor resolver
-		// service so that it shows up in the editors picker
-		this.registerDefaultEditor();
 	}
 
 	private registerListeners(): void {
@@ -116,22 +102,6 @@ export class EditorService extends Disposable implements EditorServiceImpl {
 
 		// Configuration
 		this._register(this.configurationService.onDidChangeConfiguration(e => this.onConfigurationUpdated(this.configurationService.getValue<IWorkbenchEditorConfiguration>())));
-	}
-
-	private registerDefaultEditor(): void {
-		this._register(this.editorResolverService.registerEditor(
-			'*',
-			{
-				id: DEFAULT_EDITOR_ASSOCIATION.id,
-				label: DEFAULT_EDITOR_ASSOCIATION.displayName,
-				detail: DEFAULT_EDITOR_ASSOCIATION.providerDisplayName,
-				priority: RegisteredEditorPriority.builtin
-			},
-			{},
-			editor => ({ editor: this.createEditorInput(editor) }),
-			untitledEditor => ({ editor: this.createEditorInput(untitledEditor) }),
-			diffEditor => ({ editor: this.createEditorInput(diffEditor) })
-		));
 	}
 
 	//#region Editor & group event handlers
@@ -559,9 +529,9 @@ export class EditorService extends Disposable implements EditorServiceImpl {
 			}
 		}
 
-		// Override is disabled or did not apply
+		// Override is disabled or did not apply: fallback to default
 		if (!typedEditor) {
-			typedEditor = isEditorInput(editor) ? editor : this.createEditorInput(editor);
+			typedEditor = isEditorInput(editor) ? editor : this.textEditorService.createTextEditor(editor);
 		}
 
 		// If group still isn't defined because of a disabled override we resolve it
@@ -618,9 +588,9 @@ export class EditorService extends Disposable implements EditorServiceImpl {
 				}
 			}
 
-			// Override is disabled or did not apply
+			// Override is disabled or did not apply: fallback to default
 			if (!typedEditor) {
-				typedEditor = isEditorInputWithOptions(editor) ? editor : { editor: this.createEditorInput(editor), options: editor.options };
+				typedEditor = isEditorInputWithOptions(editor) ? editor : { editor: this.textEditorService.createTextEditor(editor), options: editor.options };
 			}
 
 			// If group still isn't defined because of a disabled override we resolve it
@@ -863,11 +833,11 @@ export class EditorService extends Disposable implements EditorServiceImpl {
 				}
 			}
 
-			// Override is disabled or did not apply
+			// Override is disabled or did not apply: fallback to default
 			if (!typedReplacement) {
 				typedReplacement = {
 					editor: replacement.editor,
-					replacement: isEditorReplacement(replacement) ? replacement.replacement : this.createEditorInput(replacement.replacement),
+					replacement: isEditorReplacement(replacement) ? replacement.replacement : this.textEditorService.createTextEditor(replacement.replacement),
 					options: isEditorReplacement(replacement) ? replacement.options : replacement.replacement.options,
 					forceReplaceDirty: replacement.forceReplaceDirty
 				};
@@ -877,174 +847,6 @@ export class EditorService extends Disposable implements EditorServiceImpl {
 		}
 
 		return targetGroup?.replaceEditors(typedReplacements);
-	}
-
-	//#endregion
-
-	//#region createEditorInput()
-
-	private readonly editorInputCache = new ResourceMap<CachedEditorInput>();
-
-	createEditorInput(input: IEditorInput | IUntypedEditorInput): EditorInput;
-	createEditorInput(input: IUntypedFileEditorInput): IFileEditorInput;
-	createEditorInput(input: IEditorInput | IUntypedEditorInput | IUntypedFileEditorInput): EditorInput | IFileEditorInput {
-
-		// Typed Editor Input Support (EditorInput)
-		if (input instanceof EditorInput) {
-			return input;
-		}
-
-		// Diff Editor Support
-		if (isResourceDiffEditorInput(input)) {
-			const original = this.createEditorInput({ ...input.original });
-			const modified = this.createEditorInput({ ...input.modified });
-
-			return this.instantiationService.createInstance(DiffEditorInput, input.label, input.description, original, modified, undefined);
-		}
-
-		// Side by Side Editor Support
-		if (isResourceSideBySideEditorInput(input)) {
-			const primary = this.createEditorInput({ ...input.primary });
-			const secondary = this.createEditorInput({ ...input.secondary });
-
-			return new SideBySideEditorInput(input.label, input.description, secondary, primary);
-		}
-
-		// Untitled file support
-		const untitledInput = input as IUntitledTextResourceEditorInput;
-		if (untitledInput.forceUntitled || !untitledInput.resource || (untitledInput.resource.scheme === Schemas.untitled)) {
-			const untitledOptions = {
-				mode: untitledInput.mode,
-				initialValue: untitledInput.contents,
-				encoding: untitledInput.encoding
-			};
-
-			// Untitled resource: use as hint for an existing untitled editor
-			let untitledModel: IUntitledTextEditorModel;
-			if (untitledInput.resource?.scheme === Schemas.untitled) {
-				untitledModel = this.untitledTextEditorService.create({ untitledResource: untitledInput.resource, ...untitledOptions });
-			}
-
-			// Other resource: use as hint for associated filepath
-			else {
-				untitledModel = this.untitledTextEditorService.create({ associatedResource: untitledInput.resource, ...untitledOptions });
-			}
-
-			return this.createOrGetCached(untitledModel.resource, () => {
-
-				// Factory function for new untitled editor
-				const input = this.instantiationService.createInstance(UntitledTextEditorInput, untitledModel);
-
-				// We dispose the untitled model once the editor
-				// is being disposed. Even though we may have not
-				// created the model initially, the lifecycle for
-				// untitled is tightly coupled with the editor
-				// lifecycle for now.
-				Event.once(input.onWillDispose)(() => untitledModel.dispose());
-
-				return input;
-			}) as EditorInput;
-		}
-
-		// Text File/Resource Editor Support
-		const textResourceEditorInput = input as IUntypedFileEditorInput;
-		if (textResourceEditorInput.resource instanceof URI) {
-
-			// Derive the label from the path if not provided explicitly
-			const label = textResourceEditorInput.label || basename(textResourceEditorInput.resource);
-
-			// We keep track of the preferred resource this input is to be created
-			// with but it may be different from the canonical resource (see below)
-			const preferredResource = textResourceEditorInput.resource;
-
-			// From this moment on, only operate on the canonical resource
-			// to ensure we reduce the chance of opening the same resource
-			// with different resource forms (e.g. path casing on Windows)
-			const canonicalResource = this.uriIdentityService.asCanonicalUri(preferredResource);
-
-			return this.createOrGetCached(canonicalResource, () => {
-
-				// File
-				if (textResourceEditorInput.forceFile || this.fileService.canHandleResource(canonicalResource)) {
-					return this.fileEditorFactory.createFileEditor(canonicalResource, preferredResource, textResourceEditorInput.label, textResourceEditorInput.description, textResourceEditorInput.encoding, textResourceEditorInput.mode, textResourceEditorInput.contents, this.instantiationService);
-				}
-
-				// Resource
-				return this.instantiationService.createInstance(TextResourceEditorInput, canonicalResource, textResourceEditorInput.label, textResourceEditorInput.description, textResourceEditorInput.mode, textResourceEditorInput.contents);
-			}, cachedInput => {
-
-				// Untitled
-				if (cachedInput instanceof UntitledTextEditorInput) {
-					return;
-				}
-
-				// Files
-				else if (!(cachedInput instanceof TextResourceEditorInput)) {
-					cachedInput.setPreferredResource(preferredResource);
-
-					if (textResourceEditorInput.label) {
-						cachedInput.setPreferredName(textResourceEditorInput.label);
-					}
-
-					if (textResourceEditorInput.description) {
-						cachedInput.setPreferredDescription(textResourceEditorInput.description);
-					}
-
-					if (textResourceEditorInput.encoding) {
-						cachedInput.setPreferredEncoding(textResourceEditorInput.encoding);
-					}
-
-					if (textResourceEditorInput.mode) {
-						cachedInput.setPreferredMode(textResourceEditorInput.mode);
-					}
-
-					if (typeof textResourceEditorInput.contents === 'string') {
-						cachedInput.setPreferredContents(textResourceEditorInput.contents);
-					}
-				}
-
-				// Resources
-				else {
-					if (label) {
-						cachedInput.setName(label);
-					}
-
-					if (textResourceEditorInput.description) {
-						cachedInput.setDescription(textResourceEditorInput.description);
-					}
-
-					if (textResourceEditorInput.mode) {
-						cachedInput.setPreferredMode(textResourceEditorInput.mode);
-					}
-
-					if (typeof textResourceEditorInput.contents === 'string') {
-						cachedInput.setPreferredContents(textResourceEditorInput.contents);
-					}
-				}
-			}) as EditorInput;
-		}
-
-		throw new Error('Unknown input type');
-	}
-
-	private createOrGetCached(resource: URI, factoryFn: () => CachedEditorInput, cachedFn?: (input: CachedEditorInput) => void): CachedEditorInput {
-
-		// Return early if already cached
-		let input = this.editorInputCache.get(resource);
-		if (input) {
-			if (cachedFn) {
-				cachedFn(input);
-			}
-
-			return input;
-		}
-
-		// Otherwise create and add to cache
-		input = factoryFn();
-		this.editorInputCache.set(resource, input);
-		Event.once(input.onWillDispose)(() => this.editorInputCache.delete(resource));
-
-		return input;
 	}
 
 	//#endregion
