+++ src/vs/workbench/services/editor/test/browser/editorService.test.ts
@@ -7,32 +7,23 @@ import * as assert from 'assert';
 import { EditorActivation, EditorResolution, IResourceEditorInput } from 'vs/platform/editor/common/editor';
 import { URI } from 'vs/base/common/uri';
 import { Event } from 'vs/base/common/event';
-import { DEFAULT_EDITOR_ASSOCIATION, EditorsOrder, IEditorInputWithOptions, IEditorPane, IResourceDiffEditorInput, IResourceSideBySideEditorInput, isEditorInputWithOptions, isResourceDiffEditorInput, isResourceSideBySideEditorInput, isUntitledResourceEditorInput, IUntitledTextResourceEditorInput, IUntypedEditorInput } from 'vs/workbench/common/editor';
+import { DEFAULT_EDITOR_ASSOCIATION, EditorsOrder, IEditorInputWithOptions, IEditorPane, IResourceDiffEditorInput, isEditorInputWithOptions, IUntitledTextResourceEditorInput, IUntypedEditorInput } from 'vs/workbench/common/editor';
 import { workbenchInstantiationService, TestServiceAccessor, registerTestEditor, TestFileEditorInput, ITestInstantiationService, registerTestResourceEditor, registerTestSideBySideEditor, createEditorPart, registerTestFileEditor, TestEditorWithOptions, TestTextFileEditor } from 'vs/workbench/test/browser/workbenchTestServices';
-import { TextResourceEditorInput } from 'vs/workbench/common/editor/textResourceEditorInput';
 import { EditorService } from 'vs/workbench/services/editor/browser/editorService';
 import { IEditorGroup, IEditorGroupsService, GroupDirection, GroupsArrangement } from 'vs/workbench/services/editor/common/editorGroupsService';
 import { EditorPart } from 'vs/workbench/browser/parts/editor/editorPart';
 import { ACTIVE_GROUP, IEditorService, PreferredGroup, SIDE_GROUP } from 'vs/workbench/services/editor/common/editorService';
 import { SyncDescriptor } from 'vs/platform/instantiation/common/descriptors';
 import { FileEditorInput } from 'vs/workbench/contrib/files/browser/editors/fileEditorInput';
-import { UntitledTextEditorInput } from 'vs/workbench/services/untitled/common/untitledTextEditorInput';
 import { timeout } from 'vs/base/common/async';
-import { toResource } from 'vs/base/test/common/utils';
-import { IFileService, FileOperationEvent, FileOperation } from 'vs/platform/files/common/files';
-import { Disposable, DisposableStore } from 'vs/base/common/lifecycle';
-import { ModesRegistry } from 'vs/editor/common/modes/modesRegistry';
-import { UntitledTextEditorModel } from 'vs/workbench/services/untitled/common/untitledTextEditorModel';
-import { NullFileSystemProvider } from 'vs/platform/files/test/common/nullFileSystemProvider';
-import { DiffEditorInput } from 'vs/workbench/common/editor/diffEditorInput';
-import { isLinux } from 'vs/base/common/platform';
+import { FileOperationEvent, FileOperation } from 'vs/platform/files/common/files';
+import { DisposableStore } from 'vs/base/common/lifecycle';
 import { MockScopableContextKeyService } from 'vs/platform/keybinding/test/common/mockKeybindingService';
 import { RegisteredEditorPriority } from 'vs/workbench/services/editor/common/editorResolverService';
 import { IWorkspaceTrustRequestService, WorkspaceTrustUriResponse } from 'vs/platform/workspace/common/workspaceTrust';
 import { TestWorkspaceTrustRequestService } from 'vs/workbench/services/workspaces/test/common/testWorkspaceTrustService';
 import { SideBySideEditorInput } from 'vs/workbench/common/editor/sideBySideEditorInput';
 import { EditorInput } from 'vs/workbench/common/editor/editorInput';
-import { ITextFileEditorModel } from 'vs/workbench/services/textfile/common/textfiles';
 import { UnknownErrorEditor } from 'vs/workbench/browser/parts/editor/editorPlaceholder';
 import { TestConfigurationService } from 'vs/platform/configuration/test/common/testConfigurationService';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
@@ -41,14 +32,6 @@ suite('EditorService', () => {
 
 	const TEST_EDITOR_ID = 'MyTestEditorForEditorService';
 	const TEST_EDITOR_INPUT_ID = 'testEditorInputForEditorService';
-
-	class FileServiceProvider extends Disposable {
-		constructor(scheme: string, @IFileService fileService: IFileService) {
-			super();
-
-			this._register(fileService.registerProvider(scheme, new NullFileSystemProvider()));
-		}
-	}
 
 	const disposables = new DisposableStore();
 
@@ -1523,191 +1506,6 @@ suite('EditorService', () => {
 		}
 	});
 
-	test('caching', function () {
-		const instantiationService = workbenchInstantiationService();
-		const service = instantiationService.createInstance(EditorService);
-
-		// Cached Input (Files)
-		const fileResource1 = toResource.call(this, '/foo/bar/cache1.js');
-		const fileEditorInput1 = service.createEditorInput({ resource: fileResource1 });
-		assert.ok(fileEditorInput1);
-
-		const fileResource2 = toResource.call(this, '/foo/bar/cache2.js');
-		const fileEditorInput2 = service.createEditorInput({ resource: fileResource2 });
-		assert.ok(fileEditorInput2);
-
-		assert.notStrictEqual(fileEditorInput1, fileEditorInput2);
-
-		const fileEditorInput1Again = service.createEditorInput({ resource: fileResource1 });
-		assert.strictEqual(fileEditorInput1Again, fileEditorInput1);
-
-		fileEditorInput1Again.dispose();
-
-		assert.ok(fileEditorInput1.isDisposed());
-
-		const fileEditorInput1AgainAndAgain = service.createEditorInput({ resource: fileResource1 });
-		assert.notStrictEqual(fileEditorInput1AgainAndAgain, fileEditorInput1);
-		assert.ok(!fileEditorInput1AgainAndAgain.isDisposed());
-
-		// Cached Input (Resource)
-		const resource1 = URI.from({ scheme: 'custom', path: '/foo/bar/cache1.js' });
-		const input1 = service.createEditorInput({ resource: resource1 });
-		assert.ok(input1);
-
-		const resource2 = URI.from({ scheme: 'custom', path: '/foo/bar/cache2.js' });
-		const input2 = service.createEditorInput({ resource: resource2 });
-		assert.ok(input2);
-
-		assert.notStrictEqual(input1, input2);
-
-		const input1Again = service.createEditorInput({ resource: resource1 });
-		assert.strictEqual(input1Again, input1);
-
-		input1Again.dispose();
-
-		assert.ok(input1.isDisposed());
-
-		const input1AgainAndAgain = service.createEditorInput({ resource: resource1 });
-		assert.notStrictEqual(input1AgainAndAgain, input1);
-		assert.ok(!input1AgainAndAgain.isDisposed());
-	});
-
-	test('createEditorInput', async function () {
-		const instantiationService = workbenchInstantiationService();
-		const service = instantiationService.createInstance(EditorService);
-
-		const mode = 'create-input-test';
-		ModesRegistry.registerLanguage({
-			id: mode,
-		});
-
-		// Untyped Input (file)
-		let input = service.createEditorInput({ resource: toResource.call(this, '/index.html'), options: { selection: { startLineNumber: 1, startColumn: 1 } } });
-		assert(input instanceof FileEditorInput);
-		let contentInput = <FileEditorInput>input;
-		assert.strictEqual(contentInput.resource.fsPath, toResource.call(this, '/index.html').fsPath);
-
-		// Untyped Input (file casing)
-		input = service.createEditorInput({ resource: toResource.call(this, '/index.html') });
-		let inputDifferentCase = service.createEditorInput({ resource: toResource.call(this, '/INDEX.html') });
-
-		if (!isLinux) {
-			assert.strictEqual(input, inputDifferentCase);
-			assert.strictEqual(input.resource?.toString(), inputDifferentCase.resource?.toString());
-		} else {
-			assert.notStrictEqual(input, inputDifferentCase);
-			assert.notStrictEqual(input.resource?.toString(), inputDifferentCase.resource?.toString());
-		}
-
-		// Typed Input
-		assert.strictEqual(service.createEditorInput(input), input);
-
-		// Untyped Input (file, encoding)
-		input = service.createEditorInput({ resource: toResource.call(this, '/index.html'), encoding: 'utf16le', options: { selection: { startLineNumber: 1, startColumn: 1 } } });
-		assert(input instanceof FileEditorInput);
-		contentInput = <FileEditorInput>input;
-		assert.strictEqual(contentInput.getPreferredEncoding(), 'utf16le');
-
-		// Untyped Input (file, mode)
-		input = service.createEditorInput({ resource: toResource.call(this, '/index.html'), mode });
-		assert(input instanceof FileEditorInput);
-		contentInput = <FileEditorInput>input;
-		assert.strictEqual(contentInput.getPreferredMode(), mode);
-		let fileModel = (await contentInput.resolve() as ITextFileEditorModel);
-		assert.strictEqual(fileModel.textEditorModel?.getModeId(), mode);
-
-		// Untyped Input (file, contents)
-		input = service.createEditorInput({ resource: toResource.call(this, '/index.html'), contents: 'My contents' });
-		assert(input instanceof FileEditorInput);
-		contentInput = <FileEditorInput>input;
-		fileModel = (await contentInput.resolve() as ITextFileEditorModel);
-		assert.strictEqual(fileModel.textEditorModel?.getValue(), 'My contents');
-		assert.strictEqual(fileModel.isDirty(), true);
-
-		// Untyped Input (file, different mode)
-		input = service.createEditorInput({ resource: toResource.call(this, '/index.html'), mode: 'text' });
-		assert(input instanceof FileEditorInput);
-		contentInput = <FileEditorInput>input;
-		assert.strictEqual(contentInput.getPreferredMode(), 'text');
-
-		// Untyped Input (untitled)
-		input = service.createEditorInput({ resource: undefined, options: { selection: { startLineNumber: 1, startColumn: 1 } } });
-		assert(input instanceof UntitledTextEditorInput);
-
-		// Untyped Input (untitled with contents)
-		let untypedInput: any = { contents: 'Hello Untitled', options: { selection: { startLineNumber: 1, startColumn: 1 } } };
-		input = service.createEditorInput(untypedInput);
-		assert.ok(isUntitledResourceEditorInput(untypedInput));
-		assert(input instanceof UntitledTextEditorInput);
-		let model = await input.resolve() as UntitledTextEditorModel;
-		assert.strictEqual(model.textEditorModel?.getValue(), 'Hello Untitled');
-
-		// Untyped Input (untitled withtoUntyped2
-		input = service.createEditorInput({ resource: undefined, mode, options: { selection: { startLineNumber: 1, startColumn: 1 } } });
-		assert(input instanceof UntitledTextEditorInput);
-		model = await input.resolve() as UntitledTextEditorModel;
-		assert.strictEqual(model.getMode(), mode);
-
-		// Untyped Input (untitled with file path)
-		input = service.createEditorInput({ resource: URI.file('/some/path.txt'), forceUntitled: true, options: { selection: { startLineNumber: 1, startColumn: 1 } } });
-		assert(input instanceof UntitledTextEditorInput);
-		assert.ok((input as UntitledTextEditorInput).model.hasAssociatedFilePath);
-
-		// Untyped Input (untitled with untitled resource)
-		untypedInput = { resource: URI.parse('untitled://Untitled-1'), forceUntitled: true, options: { selection: { startLineNumber: 1, startColumn: 1 } } };
-		assert.ok(isUntitledResourceEditorInput(untypedInput));
-		input = service.createEditorInput(untypedInput);
-		assert(input instanceof UntitledTextEditorInput);
-		assert.ok(!(input as UntitledTextEditorInput).model.hasAssociatedFilePath);
-
-		// Untyped input (untitled with custom resource, but forceUntitled)
-		untypedInput = { resource: URI.file('/fake'), forceUntitled: true };
-		assert.ok(isUntitledResourceEditorInput(untypedInput));
-		input = service.createEditorInput(untypedInput);
-		assert(input instanceof UntitledTextEditorInput);
-
-		// Untyped Input (untitled with custom resource)
-		const provider = instantiationService.createInstance(FileServiceProvider, 'untitled-custom');
-
-		input = service.createEditorInput({ resource: URI.parse('untitled-custom://some/path'), forceUntitled: true, options: { selection: { startLineNumber: 1, startColumn: 1 } } });
-		assert(input instanceof UntitledTextEditorInput);
-		assert.ok((input as UntitledTextEditorInput).model.hasAssociatedFilePath);
-
-		provider.dispose();
-
-		// Untyped Input (resource)
-		input = service.createEditorInput({ resource: URI.parse('custom:resource') });
-		assert(input instanceof TextResourceEditorInput);
-
-		// Untyped Input (diff)
-		const resourceDiffInput = {
-			modified: { resource: toResource.call(this, '/modified.html') },
-			original: { resource: toResource.call(this, '/original.html') }
-		};
-		assert.strictEqual(isResourceDiffEditorInput(resourceDiffInput), true);
-		input = service.createEditorInput(resourceDiffInput);
-		assert(input instanceof DiffEditorInput);
-		assert.strictEqual(input.original.resource?.toString(), resourceDiffInput.original.resource.toString());
-		assert.strictEqual(input.modified.resource?.toString(), resourceDiffInput.modified.resource.toString());
-		const untypedDiffInput = input.toUntyped() as IResourceDiffEditorInput;
-		assert.strictEqual(untypedDiffInput.original.resource?.toString(), resourceDiffInput.original.resource.toString());
-		assert.strictEqual(untypedDiffInput.modified.resource?.toString(), resourceDiffInput.modified.resource.toString());
-
-		// Untyped Input (side by side)
-		const sideBySideResourceInput = {
-			primary: { resource: toResource.call(this, '/primary.html') },
-			secondary: { resource: toResource.call(this, '/secondary.html') }
-		};
-		assert.strictEqual(isResourceSideBySideEditorInput(sideBySideResourceInput), true);
-		input = service.createEditorInput(sideBySideResourceInput);
-		assert(input instanceof SideBySideEditorInput);
-		assert.strictEqual(input.primary.resource?.toString(), sideBySideResourceInput.primary.resource.toString());
-		assert.strictEqual(input.secondary.resource?.toString(), sideBySideResourceInput.secondary.resource.toString());
-		const untypedSideBySideInput = input.toUntyped() as IResourceSideBySideEditorInput;
-		assert.strictEqual(untypedSideBySideInput.primary.resource?.toString(), sideBySideResourceInput.primary.resource.toString());
-		assert.strictEqual(untypedSideBySideInput.secondary.resource?.toString(), sideBySideResourceInput.secondary.resource.toString());
-	});
-
 	test('close editor does not dispose when editor opened in other group', async () => {
 		const [part, service] = await createEditorService();
 
@@ -2390,7 +2188,10 @@ suite('EditorService', () => {
 	test('editorResolverService - openEditor', async function () {
 		const [, service, accessor] = await createEditorService();
 		const editorResolverService = accessor.editorResolverService;
+		const textEditorService = accessor.textEditorService;
+
 		let editorCount = 0;
+
 		const registrationDisposable = editorResolverService.registerEditor(
 			'*.md',
 			{
@@ -2402,20 +2203,24 @@ suite('EditorService', () => {
 			{},
 			(editorInput) => {
 				editorCount++;
-				return ({ editor: service.createEditorInput(editorInput) });
+				return ({ editor: textEditorService.createTextEditor(editorInput) });
 			},
 			undefined,
-			diffEditor => ({ editor: service.createEditorInput(diffEditor) })
+			diffEditor => ({ editor: textEditorService.createTextEditor(diffEditor) })
 		);
 		assert.strictEqual(editorCount, 0);
+
 		const input1 = { resource: URI.parse('file://test/path/resource1.txt') };
 		const input2 = { resource: URI.parse('file://test/path/resource1.md') };
+
 		// Open editor input 1 and it shouln't trigger override as the glob doesn't match
 		await service.openEditor(input1);
 		assert.strictEqual(editorCount, 0);
+
 		// Open editor input 2 and it should trigger override as the glob doesn match
 		await service.openEditor(input2);
 		assert.strictEqual(editorCount, 1);
+
 		// Because we specify an override we shouldn't see it triggered even if it matches
 		await service.openEditor({ ...input2, options: { override: 'default' } });
 		assert.strictEqual(editorCount, 1);
@@ -2426,7 +2231,10 @@ suite('EditorService', () => {
 	test('editorResolverService - openEditors', async function () {
 		const [, service, accessor] = await createEditorService();
 		const editorResolverService = accessor.editorResolverService;
+		const textEditorService = accessor.textEditorService;
+
 		let editorCount = 0;
+
 		const registrationDisposable = editorResolverService.registerEditor(
 			'*.md',
 			{
@@ -2438,16 +2246,18 @@ suite('EditorService', () => {
 			{},
 			(editorInput) => {
 				editorCount++;
-				return ({ editor: service.createEditorInput(editorInput) });
+				return ({ editor: textEditorService.createTextEditor(editorInput) });
 			},
 			undefined,
-			diffEditor => ({ editor: service.createEditorInput(diffEditor) })
+			diffEditor => ({ editor: textEditorService.createTextEditor(diffEditor) })
 		);
 		assert.strictEqual(editorCount, 0);
+
 		const input1 = new TestFileEditorInput(URI.parse('file://test/path/resource1.txt'), TEST_EDITOR_INPUT_ID);
 		const input2 = new TestFileEditorInput(URI.parse('file://test/path/resource2.txt'), TEST_EDITOR_INPUT_ID);
 		const input3 = new TestFileEditorInput(URI.parse('file://test/path/resource3.md'), TEST_EDITOR_INPUT_ID);
 		const input4 = new TestFileEditorInput(URI.parse('file://test/path/resource4.md'), TEST_EDITOR_INPUT_ID);
+
 		// Open editor input 1 and it shouln't trigger override as the glob doesn't match
 		await service.openEditors([{ editor: input1 }, { editor: input2 }, { editor: input3 }, { editor: input4 }]);
 		assert.strictEqual(editorCount, 2);
@@ -2458,7 +2268,10 @@ suite('EditorService', () => {
 	test('editorResolverService - replaceEditors', async function () {
 		const [part, service, accessor] = await createEditorService();
 		const editorResolverService = accessor.editorResolverService;
+		const textEditorService = accessor.textEditorService;
+
 		let editorCount = 0;
+
 		const registrationDisposable = editorResolverService.registerEditor(
 			'*.md',
 			{
@@ -2470,21 +2283,26 @@ suite('EditorService', () => {
 			{},
 			(editorInput) => {
 				editorCount++;
-				return ({ editor: service.createEditorInput(editorInput) });
+				return ({ editor: textEditorService.createTextEditor(editorInput) });
 			},
 			undefined,
-			diffEditor => ({ editor: service.createEditorInput(diffEditor) })
+			diffEditor => ({ editor: textEditorService.createTextEditor(diffEditor) })
 		);
+
 		assert.strictEqual(editorCount, 0);
+
 		const input1 = new TestFileEditorInput(URI.parse('file://test/path/resource2.md'), TEST_EDITOR_INPUT_ID);
+
 		// Open editor input 1 and it shouldn't trigger because I've disabled the override logic
 		await service.openEditor(input1, { override: EditorResolution.DISABLED });
 		assert.strictEqual(editorCount, 0);
+
 		await service.replaceEditors([{
 			editor: input1,
 			replacement: input1,
 		}], part.activeGroup);
 		assert.strictEqual(editorCount, 1);
+
 		registrationDisposable.dispose();
 	});
